+++ TraitCraft/TraitCraft.lua
@@ -14,6 +14,10 @@ TC.Default = {
     traitTable = {},
     savedCharacterList = {},
     settings = {
+      crafterRequestee = "",
+      requestOption = false,
+      receiveOption = false,
+      deleteMatchingOnRead = false,
       autoCraftOption = false,
       autoCraftNirnhoned = false,
       showKnown = false,
@@ -34,7 +38,11 @@ TC.Default = {
         g = 0.624,
         b = 0.0
       }
-    }
+    },
+    libNamespace = {
+      LDM = {},
+      LTF = {}
+    },
 }
 
 TC.currentlyLoggedInCharId = TC.currentlyLoggedInCharId or GetCurrentCharacterId()
@@ -42,7 +50,11 @@ TC.currentlyLoggedInChar = TC.currentlyLoggedInChar or {}
 TC.bitwiseChars = TC.bitwiseChars or {}
 TC.traitIndexKey = nil
 TC.hookLock = false
-
+TC.rIndices = {}
+TC.rObjects = {}
+TC.mailInstance = nil
+TC.formatter = nil
+TC.lastRequested = {}
 
 local currentlyLoggedInCharId = TC.currentlyLoggedInCharId
 local currentlyLoggedInChar = {}
@@ -52,6 +64,13 @@ local BLACKSMITH 		= CRAFTING_TYPE_BLACKSMITHING
 local CLOTHIER 			= CRAFTING_TYPE_CLOTHIER
 local WOODWORK 			= CRAFTING_TYPE_WOODWORKING
 local JEWELRY_CRAFTING 	= CRAFTING_TYPE_JEWELRYCRAFTING
+
+local CRAFT_TOKEN = {
+  [CRAFTING_TYPE_BLACKSMITHING]       = "BS",
+  [CRAFTING_TYPE_CLOTHIER]            = "CL",
+  [CRAFTING_TYPE_WOODWORKING]         = "WW",
+  [CRAFTING_TYPE_JEWELRYCRAFTING]     = "JW"
+}
 
 TC.craftingTypeIndex = 1
 TC.researchLineIndex = 1
@@ -108,7 +127,10 @@ function TC.charBitMissing(trait, mask)
   -- Indicates that character bit needs to be set or is missing (as in the case of not researched)
   -- trait is the integer bitmask
   -- mask is the power-of-two flag for the character (e.g., 1, 2, 4, 8, ...)
-  return (trait % (mask*2)) < mask
+  if trait and mask then
+    return (trait % (mask*2)) < mask
+  end
+  return true
 end
 
 function TC.ResolveTraitDiffs()
@@ -141,7 +163,7 @@ function TC.ResolveTraitDiffs()
   end
 end
 
-function TraitCraft:GetTraitFromKey(key)
+function TC:GetTraitFromKey(key)
   traitIndex = key % 10000 % 100
   researchLineIndex = (key % 10000 - traitIndex) / 100
   craftingSkillType = (key - traitIndex - (researchLineIndex * 100)) / 10000
@@ -152,10 +174,18 @@ function TraitCraft:GetTraitFromKey(key)
 	return craftingSkillType, researchLineIndex, traitIndex
 end
 
-function TraitCraft:GetTraitStringFromKey(key)
+function TC:GetTraitStringFromKey(key)
   traitIndex = key % 10000 % 100
   researchLineIndex = (key % 10000 - traitIndex) / 100
   craftingSkillType = (key - traitIndex - (researchLineIndex * 100)) / 10000
+  local craftingName = GetCraftingSkillName(craftingSkillType)
+  local researchLineName = GetSmithingResearchLineInfo(craftingSkillType,researchLineIndex)
+  local traitType, _, known = GetSmithingResearchLineTraitInfo(craftingSkillType,researchLineIndex, traitIndex)
+  local traitName = GetString("SI_ITEMTRAITTYPE", traitType)
+	return researchLineName..": "..traitName
+end
+
+function TC:GetItemString(craftingSkillType, researchLineIndex, traitIndex)
   local craftingName = GetCraftingSkillName(craftingSkillType)
   local researchLineName = GetSmithingResearchLineInfo(craftingSkillType,researchLineIndex)
   local traitType, _, known = GetSmithingResearchLineTraitInfo(craftingSkillType,researchLineIndex, traitIndex)
@@ -208,7 +238,7 @@ local function humanizeFutureTime(targetTime)
     end
 end
 
-function TraitCraft:StatsReport()
+function TC:StatsReport()
   --Build Stats Report on Current Research
   local headers = { TC.Lang.STATS_NAME, TC.Lang.STATS_TYPE, TC.Lang.STATS_RESEARCHING, TC.Lang.STATS_FINISH }
   local widths
@@ -263,12 +293,36 @@ function TraitCraft:StatsReport()
   end
 end
 
+local function registerFormatter()
+  TC.formatter:RegisterCore()
+  TC.formatter:RegisterProtocol("proto", {delimiters = { group = ":", record = ";", item = "," }})
+end
+
+local function registerTemplates()
+  TC.mailInstance:RegisterTemplate("Requestor", {
+    recipient = "{recipient}",
+    subject   = "TRAITCRAFT:RESEARCH:V1",
+    body      = "{proto}"
+  })
+
+end
+
 --When Loaded
 local function OnAddOnLoaded(eventCode, addonName)
   if addonName ~= TC.Name then return end
 	EVENT_MANAGER:UnregisterForEvent(TC.Name, EVENT_ADD_ON_LOADED)
 
   TC.AV = ZO_SavedVars:NewAccountWide("TraitCraft_Vars", 1, nil, TC.Default)
+
+  if LibTextFormat then
+    TC.formatter = TC.formatter or LibTextFormat:New(TC.AV.libNamespace.LTF)
+    registerFormatter()
+  end
+
+  if LibDynamicMail then
+    TC.mailInstance = TC.mailInstance or LibDynamicMail:New(TC.AV.libNamespace.LDM, TC.formatter)
+    registerTemplates()
+  end
 
   TC.bitwiseChars = TC.GetCharacterBitwise()
 
@@ -307,7 +361,7 @@ local function getValueFromTable(t)
     return select(2, next(t))
 end
 
-function TraitCraft:FindTraitIndex(craftingSkillType, researchLineIndex, traitType)
+function TC:FindTraitIndex(craftingSkillType, researchLineIndex, traitType)
 	local _, _, numTraits, _ = GetSmithingResearchLineInfo(craftingSkillType, researchLineIndex)
 	for traitIndex = 1, numTraits do
 		local foundTraitType, description, _ = GetSmithingResearchLineTraitInfo(craftingSkillType, researchLineIndex, traitIndex)
@@ -318,18 +372,18 @@ function TraitCraft:FindTraitIndex(craftingSkillType, researchLineIndex, traitTy
 	return ITEM_TRAIT_TYPE_NONE
 end
 
-function TraitCraft:GetTraitKey(craftingSkillType, researchLineIndex, traitIndex)
+function TC:GetTraitKey(craftingSkillType, researchLineIndex, traitIndex)
 	if craftingSkillType == nil or researchLineIndex == nil or traitIndex == nil then return end
 	return craftingSkillType * 10000 + researchLineIndex * 100 + traitIndex
 end
 
-function TraitCraft:DoesCharacterKnowTrait(craftingSkillType, researchLineIndex, traitIndex)
+function TC:DoesCharacterKnowTrait(craftingSkillType, researchLineIndex, traitIndex)
 	local _, _, knows = GetSmithingResearchLineTraitInfo(craftingSkillType, researchLineIndex, traitIndex)
 	if knows then return true end
 	return false
 end
 
-function TraitCraft:WillCharacterKnowTrait(craftingSkillType, researchLineIndex, traitIndex)
+function TC:WillCharacterKnowTrait(craftingSkillType, researchLineIndex, traitIndex)
 	local _, _, knows = GetSmithingResearchLineTraitInfo(craftingSkillType, researchLineIndex, traitIndex)
 	if knows then return true end
 	local willKnow = GetSmithingResearchLineTraitTimes(craftingSkillType, researchLineIndex, traitIndex)
@@ -337,7 +391,7 @@ function TraitCraft:WillCharacterKnowTrait(craftingSkillType, researchLineIndex,
 	return false
 end
 
-function TraitCraft:IsResearchingTrait(craftingSkillType, researchLineIndex, traitIndex)
+function TC:IsResearchingTrait(craftingSkillType, researchLineIndex, traitIndex)
 	local _, _, knows = GetSmithingResearchLineTraitInfo(craftingSkillType, researchLineIndex, traitIndex)
 	if knows then return false end
 	local willKnow = GetSmithingResearchLineTraitTimes(craftingSkillType, researchLineIndex, traitIndex)
@@ -345,13 +399,13 @@ function TraitCraft:IsResearchingTrait(craftingSkillType, researchLineIndex, tra
 	return false
 end
 
-function TraitCraft:GetResearchTimeForTrait(craftingSkillType, researchLineIndex, traitIndex)
+function TC:GetResearchTimeForTrait(craftingSkillType, researchLineIndex, traitIndex)
 	local duration, timeRemaining = GetSmithingResearchLineTraitTimes(craftingSkillType, researchLineIndex, traitIndex)
   local whenDoneTimeStamp = GetTimeStamp() + timeRemaining
   return whenDoneTimeStamp
 end
 
-function TraitCraft:SetTraitResearching(craftingType, researchLineIndex, traitIndex)
+function TC:SetTraitResearching(craftingType, researchLineIndex, traitIndex)
   local char = TC.AV.activelyResearchingCharacters[currentlyLoggedInCharId]
   if char then
     local key = TraitCraft:GetTraitKey(craftingType, researchLineIndex, traitIndex)
@@ -363,14 +417,14 @@ function TraitCraft:SetTraitResearching(craftingType, researchLineIndex, traitIn
   end
 end
 
-function TraitCraft:StopTraitResearchingWithKey(id, key)
+function TC:StopTraitResearchingWithKey(id, key)
   local char = TC.AV.activelyResearchingCharacters[id]
   if char and char.research and next(char.research) then
     char.research[key] = nil
   end
 end
 
-function TraitCraft:SetTraitKnownOnCharIdWithKey(id, key)
+function TC:SetTraitKnownOnCharIdWithKey(id, key)
   local charBitId = TC.bitwiseChars[id]
   if key and not TC.AV.traitTable[key] then
     TC.AV.traitTable[key] = 0
@@ -381,13 +435,13 @@ function TraitCraft:SetTraitKnownOnCharIdWithKey(id, key)
   TraitCraft:StopTraitResearchingWithKey(id, key)
 end
 
-function TraitCraft:SetTraitKnown(craftingType, researchLineIndex, traitIndex)
+function TC:SetTraitKnown(craftingType, researchLineIndex, traitIndex)
   local charBitId = TC.bitwiseChars[currentlyLoggedInCharId]
   local key = TraitCraft:GetTraitKey(craftingType, researchLineIndex, traitIndex)
   if key and not TC.AV.traitTable[key] then
     TC.AV.traitTable[key] = 0
   end
-  if key then
+  if key and TC.AV.traitTable[key] and charBitId then
     if TC.charBitMissing(TC.AV.traitTable[key], charBitId) then
       TC.AV.traitTable[key] = TC.AV.traitTable[key] + charBitId
     end
@@ -395,7 +449,7 @@ function TraitCraft:SetTraitKnown(craftingType, researchLineIndex, traitIndex)
   end
 end
 
-function TraitCraft:SetTraitUnknown(craftingType, researchLineIndex, traitIndex)
+function TC:SetTraitUnknown(craftingType, researchLineIndex, traitIndex)
   local charBitId = TC.bitwiseChars[currentlyLoggedInCharId]
   local key = TraitCraft:GetTraitKey(craftingType, researchLineIndex, traitIndex)
   if key then
@@ -427,7 +481,7 @@ local function SetResearchHooks()
   EVENT_MANAGER:RegisterForEvent("TC_ResearchCanceled", EVENT_SMITHING_TRAIT_RESEARCH_CANCELED, TC.SetTraitUnknown)
 end
 
-function TraitCraft:ScanKnownTraits()
+function TC:ScanKnownTraits()
   local start = GetFrameTimeMilliseconds()
   local craftTypes = { BLACKSMITH, CLOTHIER, WOODWORK, JEWELRY_CRAFTING }
   local traitLimit = 9
@@ -484,9 +538,152 @@ function TraitCraft:ScanMaxNumResearch()
   end
 end
 
+function TC.sortKeysByValue(tbl)
+  local keys = {}
+  for k in pairs(tbl) do
+      table.insert(keys, k)
+  end
+  table.sort(keys, function(a, b)
+      if tbl[a] == tbl[b] then
+          return a < b  -- tiebreaker: smaller key first
+      else
+          return tbl[a] > tbl[b]
+      end
+  end)
+  return keys
+end
+
+function TC:ScanUnknownTraitsForCrafting(charId, craftingType, scanCallback, lastCrafted)
+  local nirnCraftTypes = { CRAFTING_TYPE_BLACKSMITHING, CRAFTING_TYPE_CLOTHIER, CRAFTING_TYPE_WOODWORKING }
+  local tempResearchTable = {
+    rCounter = {},
+    rObjects = {}
+  }
+  local char = self.AV.activelyResearchingCharacters[charId]
+  if char == nil or not char["maxSimultResearch"] then
+    d(self.Lang.LOG_INTO_CHAR)
+    return
+  end
+  local scanResults = { maxSimultResearch = char["maxSimultResearch"][craftingType] }
+  local mask = self.bitwiseChars[charId]
+  if not char then
+    d(self.Lang.LOG_INTO_CHAR)
+    return
+  end
+  local research = char.research or {}
+  local researchLineLimit = GetNumSmithingResearchLines(craftingType)
+  local traitLimit = 9
+  if not self.AV.settings.autoCraftNirnhoned and self.isValueInTable(nirnCraftTypes, craftingType) then
+    traitLimit = 8
+  end
+  local key
+  local trait
+  if not self.rIndices[charId] then
+    self.rIndices[charId] = {}
+  end
+  if not self.rObjects[charId] then
+    self.rObjects[charId] = {}
+  end
+  if not self.rIndices[charId][craftingType] then
+    for r = 1, researchLineLimit do
+      if lastCrafted == nil or (lastCrafted ~= nil and not lastCrafted[charId][craftingType][r]) then
+        for t = traitLimit, 1, -1 do
+          key = self:GetTraitKey(craftingType, r, t)
+          trait = self.AV.traitTable[key] or 0
+          if self.charBitMissing(trait, mask) and not research[key] then
+            if not tempResearchTable.rCounter[r] then
+              tempResearchTable.rCounter[r] = 0
+            end
+            tempResearchTable.rCounter[r] =  tempResearchTable.rCounter[r] + 1
+            if not tempResearchTable.rObjects[r] then
+              tempResearchTable.rObjects[r] = {}
+            end
+            table.insert(tempResearchTable.rObjects[r], t)
+          end
+        end
+      end
+    end
+    self.rIndices[charId][craftingType] = TC.sortKeysByValue(tempResearchTable.rCounter)
+    self.rObjects[charId] = tempResearchTable.rObjects
+  end
+
+    --Sort by minimum research duration
+  local traitCounter = 0
+  for i = 1, #self.rIndices[charId][craftingType] do
+    local rIndex = self.rIndices[charId][craftingType][i]
+    if not scanResults[craftingType] then
+      scanResults[craftingType] = {}
+    end
+    if not scanResults[craftingType][rIndex] then
+      scanResults[craftingType][rIndex] = {}
+    end
+    for j = 1, #self.rObjects[charId][rIndex] do
+      local tIndex = self.rObjects[charId][rIndex][j]
+      scanResults[craftingType][rIndex] = tIndex
+      traitCounter = traitCounter + 1
+    end
+    if traitCounter >= char["maxSimultResearch"][craftingType] then
+      scanCallback(scanResults)
+      return
+    end
+  end
+end
+
+function TC:ScanUnknownTraitsForRequesting()
+  local charId = GetCurrentCharacterId()
+  local craftTypes = { BLACKSMITH, CLOTHIER, WOODWORK, JEWELRY_CRAFTING }
+  if not self.lastRequested[charId] then
+    self.lastRequested[charId] = {}
+  end
+  local itemDescriptions = "|r\r\n  "
+  local sendObject = {}
+  for i = 1, #craftTypes do
+    local craftingType = craftTypes[i]
+    if not self.lastRequested[charId][craftingType] then
+      self.lastRequested[charId][craftingType] = {}
+    end
+    self:ScanUnknownTraitsForCrafting(charId, craftingType, function(scanResults)
+      local record = {}
+      for rIndex, tIndex in pairs(scanResults[craftingType]) do
+        if self.lastRequested[charId][craftingType][rIndex] == nil or not self.lastRequested[charId][craftingType][rIndex][tIndex] then
+          record = { rIndex, tIndex }
+        end
+      end
+      sendObject[#sendObject + 1] = { CRAFT_TOKEN[craftingType], record }
+    end, self.lastRequested)
+  end
+  return sendObject
+end
+
+function TC:processRequestMail()
+  TC.mailInstance:RegisterInboxEvents("Requestee", "QueueItems")
+  TC.mailInstance:RegisterInboxCallback("Requestee", "QueueItems", function(mailId)
+    if self.mailInstance:CheckMailForTemplateSubject(mailId, "Requestor", "equals") then
+      local craftCounter = 0
+      local scanResults = self.mailInstance:RetrieveActiveMailData(mailId)
+      if scanResults then
+        local scope = self.formatter.Scope({ text = scanResults.body })
+        local decodedResults = self.formatter:decodeByProtocolName("proto", scope)
+        if next(decodedResults) ~= nil then
+          craftCounter = self.autocraft:CraftFromInput(decodedResults, scanResults.senderCharacterName)
+        end
+        if craftCounter > 0 then
+          if TC.AV.settings.deleteMatchingOnRead then
+            DeleteMail(mailId)
+          end
+          d(self.Lang.MAIL_PROCESSED)
+        else
+          d(self.Lang.REQUEST_NOT_PROCESSED)
+        end
+      end
+    end
+  end)
+end
+
 local function TC_Event_Player_Activated(event, isA)
 	--Only fire once after login!
 	EVENT_MANAGER:UnregisterForEvent("TC_PLAYER_ACTIVATED", EVENT_PLAYER_ACTIVATED)
+	EVENT_MANAGER:UnregisterForEvent(TC.Name.."mailbox", EVENT_MAIL_READABLE)
 	TC.currentlyLoggedInChar = {}
 	TC.BuildMenu()
 	if TC.AV.activelyResearchingCharacters[currentlyLoggedInCharId] then
@@ -500,6 +697,9 @@ local function TC_Event_Player_Activated(event, isA)
     if next(TC.AV.allCrafterIds) then
       if TC.isValueInTable(TC.AV.allCrafterIds, currentlyLoggedInCharId) then
         TC.autocraft = TC_Autocraft:New(TC)
+        if LibDynamicMail and TC.AV.settings.receiveOption then
+          EVENT_MANAGER:RegisterForEvent(TC.Name.."mailbox", EVENT_MAIL_OPEN_MAILBOX , function(mailId) TC:processRequestMail(mailId) end )
+        end
       elseif TC.autocraft then
         TC_Autocraft:Destroy()
       end
@@ -538,6 +738,20 @@ end
 
 local function setupNoop()
   return
+end
+
+function TC:GetCommonStyles()
+	-- Courtesy of Weolo and wolfstar's TraitBuddy
+	local styles = {}
+	local STYLE_KHAJIIT = 9
+	for itemStyleIndex = 1, STYLE_KHAJIIT do
+		local itemStyleId = GetValidItemStyleId(itemStyleIndex)
+		if itemStyleId > 0 then
+			-- d(sf("Adding style %s itemStyleId %s", GetItemStyleName(itemStyleId), itemStyleId))
+			styles[itemStyleId] = true
+		end
+	end
+	return styles
 end
 
 function TC.CreateIcon(control, id, key, iconPath, r, g, b, sideFloat, firstOrientation, secondOrientation, controlName)