+++ TraitCraft/TraitCraftSettings.lua
@@ -7,6 +7,8 @@ if not LibHarvensAddonSettings then
     return
 end
 
+local protocol = nil
+
 local researcherLimit = 25
 if IsInGamepadPreferredMode() then
   researcherLimit = 5
@@ -157,6 +159,11 @@ function TC.BuildMenu()
   local panel = LAM:AddAddon(TC.Name, {
     allowDefaults = false,  -- Show "Reset to Defaults" button
     allowRefresh = false    -- Enable automatic control updates
+  })
+
+  panel:AddSetting({
+    type = LAM.ST_SECTION,
+    label = TC.Lang.CRAFTER_SETTINGS,
   })
 
   panel:AddSetting {
@@ -469,21 +476,105 @@ function TC.BuildMenu()
       ReloadUI("ingame")
     end
     }
-  --Configured
-  panel:AddSetting {
-    type = LAM.ST_SECTION,
-    label = TC.Lang.ACTIVELY_RESEARCHING,
-  }
-  --Actively Researching Characters Summary
-  panel:AddSetting {
-    type = LAM.ST_LABEL,
-    label = function()
-      return TC.CurrentActivelyResearching()
-    end
-  }
-  --Breakpoint
-  panel:AddSetting {
-    type = LAM.ST_SECTION,
-    label = " ",
-  }
+    --Configured
+    panel:AddSetting {
+      type = LAM.ST_SECTION,
+      label = TC.Lang.ACTIVELY_RESEARCHING,
+    }
+    --Actively Researching Characters Summary
+    panel:AddSetting {
+      type = LAM.ST_LABEL,
+      label = function()
+        return TC.CurrentActivelyResearching()
+      end
+    }
+    panel:AddSetting({
+      type = LAM.ST_SECTION,
+      label = TC.Lang.RESEARCH_REQUESTS,
+    })
+
+    panel:AddSetting({
+        type = LibHarvensAddonSettings.ST_BUTTON,
+        label = TC.Lang.SEND_CRAFT_REQUEST,
+        buttonText = TC.Lang.AUTOFILL_REQUEST,
+        clickHandler = function(control)
+          local bodyValues = TC:ScanUnknownTraitsForRequesting()
+          local sendObject = {
+            recipient = TC.AV.settings.crafterRequestee,
+            subject = "TRAITCRAFT:RESEARCH:V1",
+            body = bodyValues,
+            records = bodyValues
+          }
+          TC.mailInstance:PopulateCompose("Requestor", sendObject)
+        end,
+        disable = function()
+          return TC.AV.settings.requestOption == false
+        end
+    })
+
+    panel:AddSetting({
+      type = LAM.ST_CHECKBOX,
+      label = TC.Lang.ENABLE_BUTTON.." "..TC.Lang.RESEARCH_REQUESTS,
+      getFunction = function() return TC.AV.settings.requestOption end,
+      setFunction = function(var)
+        TC.AV.settings.requestOption = var
+      end
+    })
+
+    panel:AddSetting({
+      type = LAM.ST_CHECKBOX,
+      label = TC.Lang.ENABLE_BUTTON.." "..TC.Lang.FULFILL_REQUEST,
+      getFunction = function() return TC.AV.settings.receiveOption end,
+      setFunction = function(var)
+        TC.AV.settings.receiveOption = var
+      end
+    })
+
+    panel:AddSetting({
+      type = LAM.ST_CHECKBOX,
+      label = TC.Lang.DELETE_ON_PROCESS,
+      getFunction = function() return TC.AV.settings.deleteMatchingOnRead end,
+      setFunction = function(var)
+        TC.AV.settings.deleteMatchingOnRead = var
+      end,
+      disable = function()
+        return TC.AV.settings.requestOption == false
+      end
+    })
+
+    panel:AddSetting({
+      type = LAM.ST_EDIT,
+      label = TC.Lang.CRAFTER_REQUESTEE,
+      getFunction = function() return "" end,
+      setFunction = function(value) TC.AV.settings.crafterRequestee = value end,
+      default = ""
+    })
+
+    panel:AddSetting({
+      type = LAM.ST_BUTTON,
+      label = TC.Lang.ACTIVE_APPLY,
+      buttonText = TC.Lang.ACTIVE_APPLY,
+      clickHandler  = function()
+        panel:UpdateControls()
+        ReloadUI("ingame")
+      end,
+      tooltip = TC.Lang.REQUIRES_RELOAD.."; "..TC.Lang.REQUIRES_LIBRARY.."LibDynamicMail, LibTextFormat",
+      disable = function()
+        return TC.AV.settings.crafterRequestee == "" or not TC.AV.settings.requestOption
+      end
+    })
+
+    panel:AddSetting({
+      type = LAM.ST_LABEL,
+      label = function()
+          return TC.AV.settings.crafterRequestee
+      end,
+      tooltip = TC.Lang.SEND_CRAFT_REQUEST
+    })
+
+    --Breakpoint
+    panel:AddSetting {
+      type = LAM.ST_SECTION,
+      label = " ",
+    }
 end