+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2026.02.04
+- Send is working, will see about inbox processing
+- Inbox processing possibly working
+- Add option for automated deletion of processed mail
+
+## 2026.01.31
+- First stab at auto-requesting research items via mail
+
 ## 2026.01.14
 - Automate deployment